pyproject.toml
[tool.poetry]
name = "zorander-v2"
version = "2.0"
description = "Rewrite of the Zorander Bot"
authors = ["thenishantsapkota <snishant306@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
jishaku = "^2.2.0"
"discord.py" = "^1.7.3"
discord = "^1.7.3"
tortoise-orm = "0.17.4"
aerich = "0.5.3"
watchgod = "^0.7"
python-dotenv = "^0.19.0"
asyncpg = "^0.23.0"
PyNaCl = "^1.4.0"
pylint = "^2.9.6"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
